# dokumentasjon/openapi-dryrun.yml
openapi: 3.1.1
info:
  title: Simuler utbetaling
  version: "3"
  description: |
    Dryrun (simuler) en utbetaling. 
    Hvert fagsystem har sitt eget endepunkt. Alle endepunktene er asynkrone via kafka i bakgrunnen, 
    de blokkerer opp til 2 min. Etter 2 minutter vil OS/UR oppheve transaksjonen, derfor har vi også 2 minutter timeout.
    Feltet dryrun i requesten blir alltid overstyrt ved bruk av rest-endepunktet.
servers:
  - url: https://utsjekk.intern.nav.no
    description: Production
  - url: https://utsjekk.intern.dev.nav.no
    description: Development
security:
  - AzureAD: []
...
paths:
  /api/dryrun/aap:
    post:
      tags:
        - AAP
      summary: Simuler utbetaling for AAP
      description: |
        Prøvekjør (dryrun) en utbetaling for AAP.
        Det er kun arjourholdte oppdrag som blir med i en simulering.
        Ajourhold skjer neste virkedag etter at utbetalingen er sendt inn.
        Det er bare `utbetal´ (azp_name) som har tilgang til denne.
      operationId: simulerAap
      parameters:
        - $ref: '#/components/parameters/TransactionID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AapUtbetaling'
      responses:
        '200':
          description: Simulering utfort.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SimuleringV1'
                  - $ref: '#/components/schemas/SimuleringV2'
                discriminator:
                  propertyName: '@type'
                  mapping:
                    v1: '#/components/schemas/SimuleringV1'
                    v2: '#/components/schemas/SimuleringV2'
        '302':
          description: |
            Simulering er vellykket, men det er ingen posteringer aa vise.
            Mulige aarsaker:
            1) Oppdraget finnes fra for, ingen endring mellom simulert og eksisterende.
            2) Det simuleres et opphor pa et oppdrag som ikke er ajourholdt.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Info'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '408':
          $ref: '#/components/responses/Timeout'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/dryrun/dagpenger:
    post:
      tags:
        - Dagpenger
      summary: Simuler utbetaling for Dagpenger
      description: |
        Prøvekjør (dryrun) en utbetaling for Dagpenger.
        Det er kun arjourholdte oppdrag som blir med i en simulering.
        Ajourhold skjer neste virkedag etter at utbetalingen er sendt inn.
        Det er bare `dp-mellom-barken-og-veden´ (azp_name) som har tilgang til denne.
      operationId: simulerDagpenger
      parameters:
        - $ref: '#/components/parameters/TransactionID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DpUtbetaling'
      responses:
        '200':
          description: Simulering utfort.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SimuleringV1'
                  - $ref: '#/components/schemas/SimuleringV2'
                discriminator:
                  propertyName: '@type'
                  mapping:
                    v1: '#/components/schemas/SimuleringV1'
                    v2: '#/components/schemas/SimuleringV2'
        '302':
          description: |
            Simulering er vellykket, men det er ingen posteringer aa vise.
            Mulige aarsaker:
            1) Oppdraget finnes fra for, ingen endring mellom simulert og eksisterende.
            2) Det simuleres et opphor pa et oppdrag som ikke er ajourholdt.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Info'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '408':
          $ref: '#/components/responses/Timeout'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/dryrun/tilleggsstonader:
    post:
      tags:
        - Tilleggsstønader
      summary: Simuler utbetaling for Tilleggsstonader
      description: |
        Prøvekjør (dryrun) en utbetaling for Dagpenger.
        Det er kun arjourholdte oppdrag som blir med i en simulering.
        Ajourhold skjer neste virkedag etter at utbetalingen er sendt inn.
        Det er bare `tilleggsstonader-sak´ (azp_name) som har tilgang til denne.
      operationId: simulerTilleggsstonader
      parameters:
        - $ref: '#/components/parameters/TransactionID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TsDto'
      responses:
        '200':
          description: Simulering utfort.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SimuleringV1'
                  - $ref: '#/components/schemas/SimuleringV2'
                discriminator:
                  propertyName: '@type'
                  mapping:
                    v1: '#/components/schemas/SimuleringV1'
                    v2: '#/components/schemas/SimuleringV2'
        '302':
          description: |
            Simulering er vellykket, men det er ingen posteringer aa vise.
            Mulige aarsaker:
            1) Oppdraget finnes fra for, ingen endring mellom simulert og eksisterende.
            2) Det simuleres et opphor pa et oppdrag som ikke er ajourholdt.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Info'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '408':
          $ref: '#/components/responses/Timeout'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/dryrun/tiltakspenger:
    post:
      tags:
        - Tiltakspenger
      summary: Simuler utbetaling for Tiltakspenger
      description: |
        Prøvekjør (dryrun) en utbetaling for Dagpenger.
        Det er kun arjourholdte oppdrag som blir med i en simulering.
        Ajourhold skjer neste virkedag etter at utbetalingen er sendt inn.
        Det er bare `tiltakspenger-saksbehandling-api´ (azp_name) som har tilgang til denne.
      operationId: simulerTiltakspenger
      parameters:
        - $ref: '#/components/parameters/TransactionID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TpUtbetaling'
      responses:
        '200':
          description: Simulering utfort.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SimuleringV1'
                  - $ref: '#/components/schemas/SimuleringV2'
                discriminator:
                  propertyName: '@type'
                  mapping:
                    v1: '#/components/schemas/SimuleringV1'
                    v2: '#/components/schemas/SimuleringV2'
        '302':
          description: |
            Simulering er vellykket, men det er ingen posteringer aa vise.
            Mulige aarsaker:
            1) Oppdraget finnes fra for, ingen endring mellom simulert og eksisterende.
            2) Det simuleres et opphor pa et oppdrag som ikke er ajourholdt.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Info'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '408':
          $ref: '#/components/responses/Timeout'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    AzureAD:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Azure AD JWT-token. Tilgang styres av `azp_name`-claimet.
        Se beskrivelse per endepunkt for hvilken `azp_name` som kreves.
  parameters:
    TransactionID:
      name: Transaction-ID
      in: header
      required: false
      description: |
        Unik transaksjons-ID for denne foresporselen. Hvis ikke angitt genereres
        en tilfeldig UUID av serveren. Brukes til korrelering via
        `CompletableDeferred` -- duplikat-ID paa en pagaende simulering gir HTTP 423.
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: |
        Ugyldig request. Returnerer enten `ApiError` (valideringsfeil)
        eller `StatusReply` (feil fra nedstromssystem).
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/ApiError'
              - $ref: '#/components/schemas/StatusReply'
    Forbidden:
      description: |
        Ikke autorisert. Returneres naar `azp_name` ikke matcher
        den forventede klienten for endepunktet.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Timeout:
      description: |
        Timeout (120 sekunder). Simuleringen ble ikke fullfort innen
        tidsfristen. Ingen response body.
...
    InternalServerError:
      description: |
        Uventet feil. Returneres naar det asynkrone resultatet
        ikke er en kjent type.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    AapUtbetaling:
      type: object
      description: Simuleringsrequest for AAP (Arbeidsavklaringspenger).
      required:
        - sakId
        - behandlingId
        - ident
        - utbetalinger
        - vedtakstidspunktet
      properties:
        dryrun:
          type: boolean
          default: false
          description: Overstyres alltid til `true` av serveren.
        sakId:
          type: string
          description: Unik sak-ID i fagsystemet.
        behandlingId:
          type: string
          description: Unik behandling-ID i fagsystemet.
        ident:
          type: string
          description: Fodselsnummer eller D-nummer til soker.
        utbetalinger:
          type: array
          items:
            $ref: '#/components/schemas/AapUtbetalingsdag'
        vedtakstidspunktet:
          type: string
          format: date-time
          description: Tidspunkt for vedtak.
        saksbehandler:
          type: string
          nullable: true
          description: NAV-ident til saksbehandler. Standard `kelvin` hvis ikke angitt.
        beslutter:
          type: string
          nullable: true
          description: NAV-ident til beslutter. Standard `kelvin` hvis ikke angitt.
        avvent:
          $ref: '#/components/schemas/Avvent'
    AapUtbetalingsdag:
      type: object
      description: En enkelt utbetalingsdag for AAP.
      required:
        - meldeperiode
        - dato
        - sats
        - utbetaltBeløp
      properties:
        meldeperiode:
          type: string
          description: Identifikator for meldeperioden.
        dato:
          type: string
          format: date
          description: Dato for utbetalingsdagen.
        sats:
          type: integer
          minimum: 0
          description: Dagsats i hele kroner.
        utbetaltBeløp:
          type: integer
          minimum: 0
          description: Faktisk utbetalt belop for denne dagen.
    Avvent:
      type: object
      description: Avvent-informasjon for utbetalingen.
      required:
        - fom
        - tom
      properties:
        fom:
          type: string
          format: date
        tom:
          type: string
          format: date
        overføres:
          type: string
          format: date
          nullable: true
        årsak:
          $ref: '#/components/schemas/Årsak'
        feilregistrering:
          type: boolean
          default: false
    Årsak:
      type: string
      description: Aarsak for avvent.
      enum:
        - AVVENT_AVREGNING
        - AVVENT_REFUSJONSKRAV
    DpUtbetaling:
      type: object
      description: Simuleringsrequest for Dagpenger.
      required:
        - sakId
        - behandlingId
        - ident
        - utbetalinger
        - vedtakstidspunktet
      properties:
        dryrun:
          type: boolean
          default: false
          description: Overstyres alltid til `true` av serveren.
        sakId:
          type: string
          description: Unik sak-ID i fagsystemet.
        behandlingId:
          type: string
          description: Unik behandling-ID i fagsystemet.
        ident:
          type: string
          description: Fodselsnummer eller D-nummer til soker.
        utbetalinger:
          type: array
          items:
            $ref: '#/components/schemas/DpUtbetalingsdag'
        vedtakstidspunktet:
          type: string
          format: date-time
          description: Tidspunkt for vedtak.
        saksbehandler:
          type: string
          nullable: true
          description: NAV-ident til saksbehandler.
        beslutter:
          type: string
          nullable: true
          description: NAV-ident til beslutter.
    DpUtbetalingsdag:
      type: object
      description: En enkelt utbetalingsdag for Dagpenger.
      required:
        - meldeperiode
        - dato
        - sats
        - utbetaltBeløp
        - utbetalingstype
      properties:
        meldeperiode:
          type: string
          description: Identifikator for meldeperioden.
        dato:
          type: string
          format: date
          description: Dato for utbetalingsdagen.
        sats:
          type: integer
          minimum: 0
          description: Dagsats i hele kroner.
        utbetaltBeløp:
          type: integer
          minimum: 0
          description: Faktisk utbetalt belop for denne dagen.
        utbetalingstype:
          $ref: '#/components/schemas/Utbetalingstype'
...
    TsDto:
      type: object
      description: Simuleringsrequest for Tilleggsstonader.
      required:
        - sakId
        - behandlingId
        - personident
        - vedtakstidspunkt
        - periodetype
        - utbetalinger
      properties:
        dryrun:
          type: boolean
          default: false
          description: Overstyres alltid til `true` av serveren.
        sakId:
          type: string
          description: Unik sak-ID i fagsystemet.
        behandlingId:
          type: string
          description: Unik behandling-ID i fagsystemet.
        personident:
          type: string
          description: Fodselsnummer eller D-nummer til soker.
        vedtakstidspunkt:
          type: string
          format: date-time
          description: Tidspunkt for vedtak.
        periodetype:
          $ref: '#/components/schemas/Periodetype'
        saksbehandler:
          type: string
          nullable: true
          description: NAV-ident til saksbehandler.
        beslutter:
          type: string
          nullable: true
          description: NAV-ident til beslutter.
        utbetalinger:
          type: array
          items:
            $ref: '#/components/schemas/TsUtbetaling'
    TsUtbetaling:
      type: object
      description: En utbetaling innenfor Tilleggsstonader.
      required:
        - id
        - stønad
        - perioder
      properties:
        id:
          type: string
          format: uuid
          description: Unik ID for denne utbetalingen.
        stønad:
          $ref: '#/components/schemas/StønadTypeTilleggsstønader'
        perioder:
          type: array
          description: Tom liste betyr sletting (opphor).
          items:
            $ref: '#/components/schemas/TsPeriode'
        brukFagområdeTillst:
          type: boolean
          default: false
          description: |
            Bruk fagomraade TILLST i stedet for spesifikke fagomraader.
            Kun tillatt for et utvalg av stonadstyper.
...
    Periodetype:
      type: string
      description: Satstype for utbetalingsperioder.
      enum:
        - DAG
        - UKEDAG
        - MND
        - EN_GANG
    TpUtbetaling:
      type: object
      description: |
        Simuleringsrequest for Tiltakspenger. Sendes som en liste (array)
        i request-body.
      required:
        - sakId
        - behandlingId
        - personident
        - vedtakstidspunkt
        - perioder
      properties:
        sakId:
          type: string
          description: Unik sak-ID i fagsystemet.
        behandlingId:
          type: string
          description: Unik behandling-ID i fagsystemet.
        dryrun:
          type: boolean
          default: false
          description: Overstyres alltid til `true` av serveren.
        personident:
          type: string
          description: Fodselsnummer eller D-nummer til soker.
        vedtakstidspunkt:
          type: string
          format: date-time
          description: Tidspunkt for vedtak.
        perioder:
          type: array
          items:
            $ref: '#/components/schemas/TpPeriode'
        saksbehandler:
          type: string
          nullable: true
          description: NAV-ident til saksbehandler.
        beslutter:
          type: string
          nullable: true
          description: NAV-ident til beslutter.
...
    SimuleringV1:
      type: object
      description: |
        Simuleringsresultat v1. Inneholder oppsummeringer per periode og
        detaljerte posteringer.
      required:
        - '@type'
        - oppsummeringer
        - detaljer
      properties:
        '@type':
          type: string
          enum:
            - v1
        oppsummeringer:
          type: array
          items:
            $ref: '#/components/schemas/OppsummeringForPeriode'
        detaljer:
          $ref: '#/components/schemas/SimuleringDetaljer'
    OppsummeringForPeriode:
      type: object
      description: Oppsummering av simulering for en periode.
      required:
        - fom
        - tom
        - tidligereUtbetalt
        - nyUtbetaling
        - totalEtterbetaling
        - totalFeilutbetaling
      properties:
        fom:
          type: string
          format: date
        tom:
          type: string
          format: date
        tidligereUtbetalt:
          type: integer
          description: Tidligere utbetalt belop.
        nyUtbetaling:
          type: integer
          description: Nytt utbetalingsbelop.
        totalEtterbetaling:
          type: integer
          description: Total etterbetaling.
        totalFeilutbetaling:
          type: integer
          description: Total feilutbetaling.
    SimuleringDetaljer:
      type: object
      description: Detaljert simuleringsresultat.
      required:
        - gjelderId
        - datoBeregnet
        - totalBeløp
        - perioder
      properties:
        gjelderId:
          type: string
          description: Person-ID simuleringen gjelder for.
        datoBeregnet:
          type: string
          format: date
          description: Dato simuleringen ble beregnet.
        totalBeløp:
          type: integer
          description: Totalt belop.
        perioder:
          type: array
          items:
            $ref: '#/components/schemas/SimuleringV1Periode'
    SimuleringV1Periode:
      type: object
      description: En periode i det detaljerte simuleringsresultatet.
      required:
        - fom
        - tom
        - posteringer
      properties:
        fom:
          type: string
          format: date
        tom:
          type: string
          format: date
        posteringer:
          type: array
          items:
            $ref: '#/components/schemas/SimuleringV1Postering'
...
    SimuleringV2:
      type: object
      description: |
        Simuleringsresultat v2. Gruppert per periode med utbetalinger og posteringer.
      required:
        - '@type'
        - perioder
      properties:
        '@type':
          type: string
          enum:
            - v2
        perioder:
          type: array
          description: Kan vaere tom liste hvis opphor av noe som ikke er effektuert enna.
          items:
            $ref: '#/components/schemas/Simuleringsperiode'
    Simuleringsperiode:
      type: object
      description: En periode i v2-simuleringsresultatet.
      required:
        - fom
        - tom
        - utbetalinger
      properties:
        fom:
          type: string
          format: date
        tom:
          type: string
          format: date
        utbetalinger:
          type: array
          items:
            $ref: '#/components/schemas/SimulertUtbetaling'
...
    Info:
      type: object
      description: |
        Informasjonsrespons for simuleringer uten posteringer aa vise.
        Returneres med HTTP 302.
      required:
        - '@type'
        - status
        - fagsystem
        - message
      properties:
        '@type':
          type: string
          enum:
            - info
        status:
          $ref: '#/components/schemas/InfoStatus'
        fagsystem:
          $ref: '#/components/schemas/Fagsystem'
        message:
          type: string
          description: Forklaring paa hvorfor det ikke er posteringer aa vise.
    InfoStatus:
      type: string
      description: Status for Info-respons.
      enum:
        - OK_UTEN_ENDRING

    # ──────────────────────────────────────────────
    #  Felles schemas
    # ──────────────────────────────────────────────
    Fagsystem:
      type: string
      description: Fagsystem-enum.
      enum:
        - DAGPENGER
        - TILTAKSPENGER
        - TILLEGGSSTØNADER
        - TILLSTPB
        - TILLSTLM
        - TILLSTBO
        - TILLSTDR
        - TILLSTRS
        - TILLSTRO
        - TILLSTRA
        - TILLSTFL
        - AAP
        - HISTORISK
    StatusReply:
      type: object
      description: |
        Statusrespons fra asynkron behandling. Returneres med HTTP 400
        naar nedstromssystemer rapporterer feil.
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/Status'
        detaljer:
          $ref: '#/components/schemas/Detaljer'
        error:
          $ref: '#/components/schemas/ApiError'
...
    ApiError:
      type: object
      description: Feilrespons fra serveren.
      required:
        - statusCode
        - msg
      properties:
        statusCode:
          type: integer
          description: HTTP-statuskode.
        msg:
          type: string
          description: Feilmelding.
        doc:
          type: string
          nullable: true
          description: Lenke til dokumentasjon for feilen.
          example: https://helved-docs.ansatt.dev.nav.no
        system:
          $ref: '#/components/schemas/System'
    System:
      type: string
      description: Hvilket system feilen oppstod i.
      enum:
        - HELVED
        - OSUR

    # ──────────────────────────────────────────────
    #  Stonadstyper per fagsystem
    # ──────────────────────────────────────────────
...